Reuse duplicated code.
